Refatora código para colocar lógica no arquivo App.js
