+++ src/App.js
@@ -1,5 +1,11 @@
-import React, {useState, useEffect } from 'react'
+import React, {useState, useEffect, useCallback } from 'react'
 import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom'
+
+import OrganizationsIcon from './assets/organizations.svg'
+import activitiesIcon from './assets/activities.svg'
+import QuestionsIcon from './assets/questions.svg'
+import Hours from './assets/hours.svg'
+
 import Horas from './views/Horas'
 import Atividades from './views/Atividades'
 import Organizacoes from './views/Organizacoes'
@@ -7,12 +13,24 @@ import Duvidas from './views/Duvidas'
 import Perfil from './views/Perfil'
 import Sidebar from './components/Sidebar'
 import { Header }from './components/Header'
+
 import './styles/reset.css'
 import './styles/root.css'
 import './styles/app.css'
 
 export const App = ({ children }) => {
     const [isMobile, setIsMobile] = useState(window.innerWidth < 1250)
+    const [pathname, setPathname] = useState(window.location.pathname)
+    const pages = [
+        { name: 'horas', text: 'Minhas Horas', icon: Hours, path: '/horas' },
+        { name: 'atividades', text: 'Minhas Atividades', icon: activitiesIcon, path: '/atividades' },
+        { name: 'orgs', text: 'Organizações', icon: OrganizationsIcon, path: '/orgs' },
+        { name: 'duvidas', text: 'Dúvidas', icon: QuestionsIcon, path: '/duvidas' },
+    ]
+    
+    const handlePathname = useCallback(() => {
+        setPathname(window.location.pathname)
+    }, [setPathname])
 
     useEffect(() => {
         window.addEventListener('resize', () => {
@@ -23,10 +41,11 @@ export const App = ({ children }) => {
         }, false)
     }, [isMobile])
 
+    const props = { pages, pathname, handlePathname }
     return (
         <div className={`app`}>
             <BrowserRouter>
-                {isMobile ? <Header/> : <Sidebar/>}
+                { isMobile ? <Header props={props}/> : <Sidebar props={props}/> }
                 <div className={`${isMobile ? 'mobile-body' : 'body'}`}>
                     <Switch>
                         <Route path='/' exact component={Horas} />
