+++ src/components/NavbarLink/index.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
+
 import style from './style.module.css'
 
 const NavbarLink = ({icon, text, path, pathname, retracted}) => {
