+++ src/components/Profile/index.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
+
 import style from './style.module.css'
 
 export const Profile = ({photo, name, retracted}) => {
