+++ src/components/Sidebar/index.js
@@ -1,34 +1,24 @@
-import React, { useState, useCallback, useEffect } from 'react'
+import React, { useState, useCallback } from 'react'
+
 import { ReactComponent as Retract } from '../../assets/retract.svg'
-import OrganizationsIcon from '../../assets/organizations.svg'
-import activitiesIcon from '../../assets/activities.svg'
-import QuestionsIcon from '../../assets/questions.svg'
-import {ReactComponent as HocoLogo} from '../../assets/LogoHoco.svg'
-import Hours from '../../assets/hours.svg'
+import { ReactComponent as HocoLogo } from '../../assets/LogoHoco.svg'
 import Photo from '../../assets/eu.jpeg'
-import style from './style.module.css'
+
 import NavbarLink from '../NavbarLink'
 import { Profile } from '../Profile'
 import { Logo } from '../Logo'
 
-const Sidebar = () => {
-    const [pathname, setPathname] = useState(window.location.pathname)
+import style from './style.module.css'
+
+const Sidebar = ({ props }) => {
+    const {pages, pathname, handlePathname} = props
+
     const [retracted, setRetracted] = useState(false)
     const [currentOpacity, setOpacity] = useState('1')
-    const pages = [
-        { name: 'horas', text: 'Minhas Horas', icon: Hours, path: '/horas' },
-        { name: 'atividades', text: 'Minhas Atividades', icon: activitiesIcon, path: '/atividades' },
-        { name: 'orgs', text: 'Organizações', icon: OrganizationsIcon, path: '/orgs' },
-        { name: 'duvidas', text: 'Dúvidas', icon: QuestionsIcon, path: '/duvidas' },
-    ]
 
     function sleep(ms) {
         return new Promise(resolve => setTimeout(resolve, ms));
     }
-
-    const handlePathname = useCallback(() => {
-        setPathname(window.location.pathname)
-    }, [setPathname])
 
     const handleRetract = useCallback(async () => {
         setRetracted(!retracted)
@@ -47,7 +37,7 @@ const Sidebar = () => {
                             className={`${style.retract}`}
                             style={{ opacity: currentOpacity }} />
                         <div className={style.logo}>
-                            <Logo Logo={HocoLogo}/>
+                            <Logo Logo={HocoLogo} />
                         </div>
                     </div>
